% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-structure.R
\name{calculate_index_of_start_and_stop_lines}
\alias{calculate_index_of_start_and_stop_lines}
\title{Calculate the index of lines where functions start and stop.}
\usage{
calculate_index_of_start_and_stop_lines(lines, n_level = 0L)
}
\arguments{
\item{lines}{An array of characters, representing the lines of code}

\item{n_level}{An integer, 0 if a definition of a function inside a function
is forbidden.}
}
\value{
A list, the start and stop indicies
}
\description{
Calculate the index of lines where functions start and stop.
}
