% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-functions.R
\name{calculate_number_of_open_braces}
\alias{calculate_number_of_open_braces}
\title{Create array of indicies}
\usage{
calculate_number_of_open_braces(lines_of_code)
}
\arguments{
\item{lines_of_code}{An array, containing the full lines of code as character}
}
\value{
An array
}
\description{
The value at each position increases (decreases) by one if a brace is opened (closed)
in the corresponding line of the code.
}
