% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphNEL.R
\name{create_graphNEL_object}
\alias{create_graphNEL_object}
\title{Create a graphNEL object}
\usage{
create_graphNEL_object(my_structure)
}
\arguments{
\item{my_structure}{A list, the functional structure calculated in
create_list_of_functional_structure().}
}
\value{
A graphNEL object
}
\description{
Create a graphNEL object from the functional structure.
}
\examples{
my_structure <- create_list_of_functional_structure(filename, path)
my_graph <- create_graphNEL_object(my_structure)
attrs <- list(node=list(fontsize = 8, shape="ellipse", fixedsize=FALSE,
                        label="foo"),
              edge=list(minlen = 2),
              graph=list(rankdir="LR"))
Rgraphviz::plot(my_graph, attrs=attrs)
Rgraphviz::plot(my_graph, "neato", attrs=attrs)
Rgraphviz::plot(my_graph, "twopi", attrs=attrs)
}
