% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-functions.R
\name{create_list_of_functional_structure}
\alias{create_list_of_functional_structure}
\title{Create a list of the functional structure}
\usage{
create_list_of_functional_structure(filename, path)
}
\arguments{
\item{filename}{A character, the file name}

\item{path}{A character, the path to the file}
}
\value{
A list
}
\description{
The list contains all important information (the start index, the stop index,
the names of dependent functions, the arguments, the corresponding defaults,
the file name and the path to the file).
}
\examples{
my_structure <- create_list_of_functional_structure(filename, path)
}
