% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetwork.R
\name{get_edges_from_structure}
\alias{get_edges_from_structure}
\title{Caluclate edges}
\usage{
get_edges_from_structure(my_structure)
}
\arguments{
\item{my_structure}{A list, the functional structure calculated in
create_list_of_functional_structure().}
}
\value{
A data.frame, the edges
}
\description{
Caluclate edges
}
