% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetwork.R
\name{get_parents_of}
\alias{get_parents_of}
\title{Retrieve the parents of a node}
\usage{
get_parents_of(node, my_structure)
}
\arguments{
\item{node}{A data.frame, the nodes calculated in get_nodes_from_structure().}

\item{my_structure}{A list, the functional structure calculated in
create_list_of_functional_structure().}
}
\value{
An array, the parent nodes
}
\description{
Each node represents a function defined in the code.
}
