% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-structure.R
\name{get_variables_defaults}
\alias{get_variables_defaults}
\title{Get the variables and the corresponding default values}
\usage{
get_variables_defaults(list_of_function_arguments, i)
}
\arguments{
\item{list_of_function_arguments}{An array of characters, the functions' names}

\item{i}{An integer, the index}
}
\value{
A list containing the variables names and the default values.
}
\description{
Return the values for the i-th function with respect to the "array of function names".
}
