% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_code.R
\name{validate_code}
\alias{validate_code}
\title{Analyse code structure for critical issues}
\usage{
validate_code(functional_structure)
}
\arguments{
\item{functional_structure}{A list, the functional structure calculate from
create_list_of_functional_structure()}
}
\value{
flag An integer. 0 = everything ok, 1 = warnings, -1 = bugs.
}
\description{
This function analyses the code structure using the concept of graphs. At the
moment the function is a beta version
}
\examples{
functional_structure <- badhacker::create_list_of_functional_structure(filenames, paths)
validate_code(functional_structure)
}
