% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetwork.R
\name{visualize_functional_structure}
\alias{visualize_functional_structure}
\title{Display a graph using visNetwork of the functional structure.}
\usage{
visualize_functional_structure(my_structure)
}
\arguments{
\item{my_structure}{A list, the functional structure calculated in
create_list_of_functional_structure().}
}
\value{
A list
}
\description{
Display a graph using visNetwork of the functional structure.
}
\examples{
my_structure <- create_list_of_functional_structure(filename, path)
visualize_functional_structure(my_structure)
}
